Recreate exp.dataancestors table upon upgrade

+++ experiment/src/org/labkey/experiment/api/ClosureQueryHelper.java
@@ -49,7 +49,7 @@ import static org.labkey.api.exp.api.ExpProtocol.ApplicationType.ExperimentRunOu
 
 public class ClosureQueryHelper
 {
-    private final static Logger logger = LogHelper.getLogger(ClosureQueryHelper.class, "Sample and data class object ancestor data computation.");
+    private final static Logger LOG = LogHelper.getLogger(ClosureQueryHelper.class, "Sample and data class object ancestor data computation.");
     final static String CONCEPT_URI = "http://www.labkey.org/types#ancestorLookup";
 
     // N.B., This should be twice the number of generations we expect as a maximum number of ancestors due to the run nodes.
@@ -309,7 +309,7 @@ public class ClosureQueryHelper
             SQLFragment selectInto = selectIntoTempTableSql(getScope().getSqlDialect(), from, tableNameSql);
             selectInto.addTempToken(ref);
             int count = new SqlExecutor(getScope()).execute(selectInto);
-            logger.debug("Selected {} rows into temp.{} for recompute of {} ancestors.", count, tempTableName, isSampleType ? "sample" : "data");
+            LOG.debug("Selected {} rows into temp.{} for recompute of {} ancestors.", count, tempTableName, isSampleType ? "sample" : "data");
 
             SQLFragment upsert;
             TableInfo tInfo = isSampleType ? ExperimentServiceImpl.get().getTinfoMaterialAncestors() : ExperimentServiceImpl.get().getTinfoDataAncestors();
@@ -418,22 +418,31 @@ public class ClosureQueryHelper
         logger.info("Finished populating exp.dataAncestors");
     }
 
-    public static void truncateAndRecreate()
+    public static void truncateAndRecreate(@Nullable Logger logger)
     {
-        truncateAndRecreate(logger);
+        truncateAndRecreate(logger, true, true);
     }
 
-    public static void truncateAndRecreate(Logger logger)
+    public static void truncateAndRecreate(@Nullable Logger logger, boolean materialAncestors, boolean dataAncestors)
     {
         try (DbScope.Transaction transaction = ExperimentService.get().getSchema().getScope().ensureTransaction())
         {
-            TableInfo tInfo = ExperimentServiceImpl.get().getTinfoMaterialAncestors();
-            new SqlExecutor(tInfo.getSchema()).execute("TRUNCATE TABLE " + tInfo);
-            ClosureQueryHelper.populateMaterialAncestors(logger);
+            if (logger == null)
+                logger = LOG;
 
-            tInfo = ExperimentServiceImpl.get().getTinfoDataAncestors();
-            new SqlExecutor(tInfo.getSchema()).execute("TRUNCATE TABLE " + tInfo);
-            ClosureQueryHelper.populateDataAncestors(logger);
+            if (materialAncestors)
+            {
+                TableInfo tInfo = ExperimentServiceImpl.get().getTinfoMaterialAncestors();
+                new SqlExecutor(tInfo.getSchema()).execute("TRUNCATE TABLE " + tInfo);
+                ClosureQueryHelper.populateMaterialAncestors(logger);
+            }
+
+            if (dataAncestors)
+            {
+                TableInfo tInfo = ExperimentServiceImpl.get().getTinfoDataAncestors();
+                new SqlExecutor(tInfo.getSchema()).execute("TRUNCATE TABLE " + tInfo);
+                ClosureQueryHelper.populateDataAncestors(logger);
+            }
 
             transaction.commit();
         }
@@ -456,7 +465,7 @@ public class ClosureQueryHelper
                 // complete hack to get SQLServer to not make RowId an identity column in the target table so the subsequent insert will work without complaint
                 selectIntoSql.append(" UNION ALL SELECT RowId, ObjectId, 'x' AS ObjectType FROM " ).append(isSampleType ? "exp.material" : "exp.data").append(" WHERE 1 <> 1");
             int numSeeds = new SqlExecutor(getScope()).execute(selectIntoSql);
-            logger.debug("Added {} seed {} rows to temp.{}", numSeeds, isSampleType ? "sample" : "data", familyTableName);
+            LOG.debug("Added {} seed {} rows to temp.{}", numSeeds, isSampleType ? "sample" : "data", familyTableName);
             // if we didn't actually insert any items into the table, there's nothing more to be done
             if (numSeeds == 0)
                 return;
@@ -472,7 +481,7 @@ public class ClosureQueryHelper
             descendants.append("INSERT INTO temp.").append(tableNameSql)
                     .append(" (RowId, ObjectId, ObjectType) ").append(descendantClosureSelectSql);
             int numRows = new SqlExecutor(getScope()).execute(descendants);
-            logger.debug("Added {} {} descendant rows for {}", numRows, isSampleType ? "sample" : "data", familyTableName);
+            LOG.debug("Added {} {} descendant rows for {}", numRows, isSampleType ? "sample" : "data", familyTableName);
 
             // recompute the ancestors for the seed ids and the descendants
             incrementalRecomputeFromTempTable(tableNameSql, isSampleType);
